+++ src/game/game/game.c
@@ -107,7 +107,7 @@ static int32_t Game_Control(int32_t nframes, GAMEFLOW_LEVEL_TYPE level_type)
         Effect_Control();
 
         Lara_Control();
-        Lara_Hair_Control(false);
+        Lara_Hair_Control();
 
         Camera_Update();
         Effect_RunActiveFlipEffect();