+++ src/game/lara/lara_hair.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include <stdbool.h>
+#include "global/types.h"
 
 void Lara_Hair_Initialise(void);
-void Lara_Hair_Control(bool in_cutscene);
+void Lara_Hair_SetLaraType(GAME_OBJECT_ID lara_type);
+void Lara_Hair_Control(void);
 void Lara_Hair_Draw(void);